# case-tracking/case-tracking-service-with-rust/openapi.yaml
openapi: 3.1.0
info:
  title: Case Tracking API (Loco edition)
  version: 0.1.0
  description: >
    JSON API for tracking the physical location of paper case-note folders
    across NHS hospital cabinets. The tracker owns no domain tables; every
    entity is proxied from one of five upstream Main-X-Services. See
    `spec/index.md` for the full specification. This document is the
    machine-readable form of `spec/api-contract.md` + `spec/routes.md`.
servers:
  - url: http://localhost:5150
    description: Local dev (JSON API). In dev the SvelteKit app proxies /api here.
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/volumes:
    get:
      tags: [volumes]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: nhs_number, in: query, schema: { type: string } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeList' } } }
          description: OK
        '503': { $ref: '#/components/responses/Upstream' }
    post:
      tags: [volumes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nhs_number, title]
              properties:
                nhs_number: { type: string }
                title: { type: string }
                cabinet_id: { type: [string, 'null'], format: uuid }
      responses:
        '201':
          description: Created
          headers: { Location: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Volume' } } }
        '422': { $ref: '#/components/responses/Validation' }
        '503': { $ref: '#/components/responses/Upstream' }
  /api/volumes/{id}:
    get:
      tags: [volumes]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [volumes]
      summary: Rename a volume.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties: { title: { type: string } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
  /api/volumes/{id}/folders:
    post:
      tags: [volumes]
      summary: Add a folder to the volume (same patient).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder_id]
              properties: { folder_id: { type: string, format: uuid } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
  /api/volumes/{id}/folders/{folder_id}:
    delete:
      tags: [volumes]
      summary: Remove a folder from the volume.
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: folder_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
  /api/volumes/{id}/move:
    post:
      tags: [volumes]
      summary: Move the whole volume — relocates every member folder, one move event each.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_cabinet_id: { type: [string, 'null'], format: uuid }
                worker_id: { type: [string, 'null'], format: uuid }
                moved_by: { type: [string, 'null'] }
                reason: { type: [string, 'null'] }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/VolumeShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
        '503': { $ref: '#/components/responses/Upstream' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cts_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Upstream:
      description: An upstream Main-X-Service is unreachable.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Validation:
      description: Validation failure — each failing field is keyed.
      content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties: { type: string }
...
    Folder:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        patient_id: { type: string, format: uuid }
        nhs_number: { type: string }
        patient_name: { type: string }
        cabinet_id: { type: [string, 'null'], format: uuid }
        cabinet_label: { type: string }
        status: { type: string, enum: [in-cabinet, in-transit] }
        last_moved_at: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }
        volume_id: { type: [string, 'null'], format: uuid }
        volume_title: { type: [string, 'null'] }
...
    Move:
      type: object
      properties:
        id: { type: string, format: uuid }
        folder_id: { type: string, format: uuid }
        folder_title: { type: string }
        patient_id: { type: string, format: uuid }
        nhs_number: { type: string }
        patient_name: { type: string }
        from_cabinet_id: { type: [string, 'null'], format: uuid }
        from_cabinet_label: { type: string }
        to_cabinet_id: { type: [string, 'null'], format: uuid }
        to_cabinet_label: { type: string }
        worker_id: { type: [string, 'null'], format: uuid }
        moved_by: { type: string }
        worker_role: { type: [string, 'null'] }
        moved_at: { type: string, format: date-time }
        reason: { type: [string, 'null'] }
...
    Volume:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        patient_id: { type: string, format: uuid }
        nhs_number: { type: string }
        patient_name: { type: string }
        cabinet_id: { type: [string, 'null'], format: uuid }
        cabinet_label: { type: string }
        status: { type: string, enum: [in-cabinet, in-transit] }
        folder_count: { type: integer }
    VolumeShow:
      allOf:
        - { $ref: '#/components/schemas/Volume' }
        - type: object
          properties:
            folders: { type: array, items: { $ref: '#/components/schemas/Folder' } }
            history: { type: array, items: { $ref: '#/components/schemas/Move' } }
...
    VolumeList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Volume' } }
        query: { type: string }
...